Moved all types into separate types files, updated all imports
