+++ src/app/framework/accordion/accordion.component.ts
@@ -1,5 +1,4 @@
 import { Component, OnInit } from '@angular/core';
-
 @Component({
   selector: 'app-accordion',
   templateUrl: './accordion.component.html',
