+++ src/app/framework/accordion/accordion.types.ts
@@ -1,0 +1,6 @@
+interface CollapseEvent {
+  headerTitle: string | null;
+  collapsed: boolean;
+}
+
+export { CollapseEvent };
