+++ src/app/framework/button/button.types.ts
@@ -1,0 +1,3 @@
+type ButtonColor = 'primary' | 'secondary' | 'success' | 'info' | 'warning' | 'danger' | 'outline';
+
+export { ButtonColor };
