+++ src/app/framework/pagination/pagination.types.ts
@@ -1,0 +1,3 @@
+type PageSize = 5 | 10 | 20 | 50 | 100 | 'All';
+
+export { PageSize };
