+++ src/app/framework/table/table.types.ts
@@ -1,0 +1,13 @@
+type TableColumnAlignment = 'left' | 'center' | 'right';
+
+interface TableColumn {
+  headerTitle: string;
+  textAlign: TableColumnAlignment;
+  width?: string;
+}
+
+interface TableData {
+  [key: string]: any;
+}
+
+export { TableColumn, TableData };
