+++ src/app/layout/header/header.types.ts
@@ -1,0 +1,7 @@
+interface NavLink {
+  name: string;
+  route: string;
+  matchExactRouteUrl: boolean;
+}
+
+export { NavLink };
