+++ src/app/pages/settings/components/add-character-form/add-character-form.component.ts
@@ -2,7 +2,7 @@ import { Component, EventEmitter, Input, OnChanges, OnInit, Output, SimpleChange
 import { FormControl, FormGroup } from '@angular/forms';
 import { faAsterisk, faInfoCircle } from '@fortawesome/free-solid-svg-icons';
 import { CharacterService } from 'src/app/utils/services/character.service';
-import { CharacterInfo } from '../../settings.component';
+import { CharacterInfo } from '../../settings.types';
 
 @Component({
   selector: 'app-add-character-form',
